Feature/exo 16 add endpoint for fact entity (#41)
Add endpoints
Update, delete, get and connecto fac to to character and disconnect fact from characters
refactor version entity handle -> changed to ushort
Moved dto files to new direcotory

+++ src/LoreWeave.Application/Commands/UpdateKnowRelationCommand.cs
@@ -5,4 +5,4 @@ public record UpdateKnowRelationCommand(
     Guid ToCharacterId,
     string Description,
     bool IsStrongRelation,
-    int Version);
+    ushort Version);
