+++ src/LoreWeave.Domain/Models/EntityExistence.cs
@@ -1,3 +1,3 @@
 namespace LoreWeave.Domain.Models;
 
-public record EntityExistence(bool Exists, int Version);
+public record EntityExistence(bool Exists, ushort Version);
